BiomedGPTvisulize/requirements.txt
absl==0.0
aml==0.0.0
apex==0.9.10dev
azureml==0.2.7
bitarray==3.4.0
boto3==1.38.14
botocore==1.38.14
Cython==3.0.12
deepspeed==0.16.7
editdistance==0.8.1
einops==0.8.1
fairscale==0.4.13
faiss==1.5.3
fastBPE==0.1.0
fasttext==0.9.3
flashlight==0.1.1
ftfy==6.3.1
g2p_en==2.1.0
gossip==2.4.0
h5py==3.13.0
indicnlp==0.0.1
inflect==7.5.0
iopath==0.1.10
joblib==1.5.0
kaldi_io==0.9.8
kenlm==0.3.0
librosa==0.11.0
matplotlib==3.5.3
nltk==3.9.1
npy_append_array==0.9.16
numpy==1.21.6
omegaconf==2.0.6
opencv_python==4.11.0.86
pandas==1.3.5
phonemizer==3.3.0
Pillow==9.5.0
Pillow==11.2.1
progressbar33==2.4
psutil==5.9.3
pyarrow==12.0.1
pycocoevalcap==1.2
pythainlp==5.1.2
pytorch_lightning==2.5.1.post0
pyworld==0.3.5
PyYAML==6.0.1
PyYAML==6.0.2
regex==2024.4.16
Requests==2.32.3
rouge_score==0.1.2
sacrebleu==2.4.2
sacremoses==0.1.1
scikit_learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
setuptools==65.6.3
simuleval==1.1.4
six==1.16.0
six==1.17.0
soundfile==0.13.1
spacy==3.8.5
submitit==1.5.2
subword_nmt==0.3.8
tabulate==0.9.0
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
tgt==1.5
timm==0.6.13
torch==1.13.1
torch_xla==2.7.0
torchaudio==2.7.0
torchvision==0.14.1
tqdm==4.67.1
transformers==4.51.3
Unidecode==1.4.0
wandb==0.19.11
webrtcvad==2.0.10
wget==3.2
zhconv==1.4.3
